// packages/remargin-obsidian/src/generated/types.ts
export type AuthorType =
  /*
  * An AI agent participant.
  */
  | "Agent"
  /*
  * A human participant.
  */
  | "Human";
/**
 * An acknowledgment of a comment by another participant.
 * 
 * JSON Schema:
 * {
 *   "additionalProperties": false,
 *   "properties": {
 *     "author": {
 *       "type": "string"
 *     },
 *     "ts": {
 *       "format": "date-time",
 *       "type": "string"
 *     }
 *   },
 *   "required": [
 *     "author",
 *     "ts"
 *   ],
 *   "type": "object"
 * }
 **/


export type Acknowledgment = {
  /**
 * Author who acknowledged.
 * 
**/
  author: string;
  /**
 * Timestamp of the acknowledgment.
 * 
**/
  ts: string;

};
// ...
export type ExpandedComment = {
  /**

**/
  ack: Array<Acknowledgment>;
  /**

**/
  attachments: Array<string>;
  /**
 * Author name or identifier.
 * 
**/
  author: string;
  /**
 * Whether the author is human or agent.
 * 
**/
  author_type: AuthorType;
  /**
 * Content integrity checksum.
 * 
**/
  checksum: string;
  /**
 * Comment body text.
 * 
**/
  content: string;
  /**
 * Relative path to the file this comment belongs to.
 * 
**/
  file: string;
  /**
 * Unique short identifier.
 * 
**/
  id: string;
  /**
 * 1-indexed line number in the source document.
 * 
**/
  line: number;
  /**
 * Emoji reactions mapped to lists of author IDs.
 * 
**/
  reactions: Partial<Record<string, Array<string>>>;
  /**

**/
  reply_to: string | undefined;
  /**

**/
  signature: string | undefined;
  /**

**/
  thread: string | undefined;
  /**

**/
  to: Array<string>;
  /**
 * Timestamp when the comment was created.
 * 
**/
  ts: string;

};